+++ examples/mongodbatlas_cloud_backup_snapshot_export_job/main.tf
@@ -22,20 +22,28 @@ resource "aws_s3_bucket" "test_bucket" {
   }
 }
 
-resource "mongodbatlas_cluster" "my_cluster" {
-  project_id   = var.project_id
-  name         = "MyCluster"
-  disk_size_gb = 1
+resource "mongodbatlas_advanced_cluster" "my_cluster" {
+  project_id     = var.project_id
+  name           = "MyCluster"
+  cluster_type   = "REPLICASET"
+  backup_enabled = true
 
-  provider_name               = "AWS"
-  provider_region_name        = "US_EAST_1"
-  provider_instance_size_name = "M10"
-  cloud_backup                = true
+  replication_specs {
+    region_configs {
+      priority      = 7
+      provider_name = "AWS"
+      region_name   = "US_EAST_1"
+      electable_specs {
+        instance_size = "M10"
+        node_count    = 3
+      }
+    }
+  }
 }
 
 resource "mongodbatlas_cloud_backup_snapshot" "test" {
   project_id        = var.project_id
-  cluster_name      = mongodbatlas_cluster.my_cluster.name
+  cluster_name      = mongodbatlas_advanced_cluster.my_cluster.name
   description       = "myDescription"
   retention_in_days = 1
 }
@@ -50,7 +58,7 @@ resource "mongodbatlas_cloud_backup_snapshot_export_bucket" "test" {
 
 resource "mongodbatlas_cloud_backup_snapshot_export_job" "test" {
   project_id       = var.project_id
-  cluster_name     = mongodbatlas_cluster.my_cluster.name
+  cluster_name     = mongodbatlas_advanced_cluster.my_cluster.name
   snapshot_id      = mongodbatlas_cloud_backup_snapshot.test.snapshot_id
   export_bucket_id = mongodbatlas_cloud_backup_snapshot_export_bucket.test.export_bucket_id