+++ examples/starter/atlas_cluster.tf
@@ -1,24 +1,23 @@
-resource "mongodbatlas_cluster" "cluster" {
-  project_id             = mongodbatlas_project.project.id
-  name                   = var.cluster_name
-  mongo_db_major_version = var.mongodbversion
-  cluster_type           = "REPLICASET"
+resource "mongodbatlas_advanced_cluster" "cluster" {
+  project_id     = mongodbatlas_project.project.id
+  name           = var.cluster_name
+  cluster_type   = "REPLICASET"
+  backup_enabled = true
+
   replication_specs {
-    num_shards = 1
-    regions_config {
-      region_name     = var.region
-      electable_nodes = 3
-      priority        = 7
-      read_only_nodes = 0
+    region_configs {
+      priority      = 7
+      provider_name = var.cloud_provider
+      region_name   = var.region
+      electable_specs {
+        instance_size = "M10"
+        node_count    = 3
+      }
     }
   }
-  # Provider Settings "block"
-  cloud_backup                 = true
-  auto_scaling_disk_gb_enabled = true
-  provider_name                = var.cloud_provider
-  provider_instance_size_name  = "M10"
 }
+
 output "connection_strings" {
-  value = mongodbatlas_cluster.cluster.connection_strings[0].standard_srv
+  value = mongodbatlas_advanced_cluster.cluster.connection_strings[0].standard_srv
 }